requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0